refactor: standardize import statements and enhance utility functions with color constants
